long type for matchSeqNum parameter

+++ protocols/valve/dota2/webapi/src/main/java/com/ibasco/agql/protocols/valve/dota2/webapi/interfaces/match/GetMatchHistoryBySequenceNum.java
@@ -28,7 +28,7 @@ import com.ibasco.agql.protocols.valve.dota2.webapi.Dota2ApiConstants;
 import com.ibasco.agql.protocols.valve.dota2.webapi.requests.Dota2MatchRequest;
 
 public class GetMatchHistoryBySequenceNum extends Dota2MatchRequest {
-    public GetMatchHistoryBySequenceNum(int apiVersion, int startSeqNum, int matchesRequested) {
+    public GetMatchHistoryBySequenceNum(int apiVersion, long startSeqNum, int matchesRequested) {
         super(Dota2ApiConstants.DOTA2_METHOD_GETMATCHHISTORYBYSEQNUM, apiVersion, null);
         urlParam("start_at_match_seq_num", startSeqNum);
         urlParam("matches_requested", matchesRequested);
